requirements.txt
langchain==0.0.242
openai==0.27.8
python-pptx==0.6.21
backoff==2.2.1
pinecone-client==2.2.2